draft idea for splitting internal bootstrap.Config
